Changes based on feedback 2
